+++ Quicksilver/Code-QuickStepEffects/QSShading.m
@@ -40,7 +40,7 @@ void QSFillRectWithGradientFromEdge(NSRect rect, NSColor *start, NSColor *end, N
 	static const CGFloat range[10] = { 0, 1, 0, 1, 0, 1, 0, 1, 0, 1 } ;
 	static const CGFunctionCallbacks callbacks = { 0, &QSColorFade, NULL } ;
 
-	float colors[8] = {
+	CGFloat colors[8] = {
 		[start redComponent] , [start greenComponent] , [start blueComponent] , [start alphaComponent] ,
 		[end redComponent] , [end greenComponent] , [end blueComponent] , [end alphaComponent] } ;
 
@@ -86,7 +86,7 @@ NSBezierPath *QSGlossClipPathForRectAndStyle(NSRect rect, QSGlossStyle style) {
 		case QSGlossControl:
 			[gloss setWindingRule:NSNonZeroWindingRule];
 
-			float radius = NSHeight(rect) /2;
+			CGFloat radius = NSHeight(rect) /2;
 			[gloss appendBezierPathWithArcWithCenter:NSMakePoint(NSMinX(rect) +radius, NSMinY(rect))
 											 radius:radius
 										 startAngle:180.0
